Feat: Lift up item and buy count states
Also create custom hook for fetching to separate
code.

The items and buy count states (and update to
buy counts curry function) was lifted up to the
Layout.jsx level and using Context API to pass
to children. This is needed because I don't
want these states to be reset upon unmounting
of the Items component (and need to share this
state with a lot of components)