Fix: Remove extra contexting for cart aside
Previously created another layout component
to hold the cart aside, but it is better to just
manually add the CartAside component to the
needed routes, instead of creating another, complex
outlet. Since we only have 1-2 pages that need
this.

Also: any child components rendered in a route
outlet will also be able to retrive the context
values.