Fix: Pass outlet context as object, not array
This is better because passing an array requires
users of the context to know the exact ordering
of the data.

Also: in general be careful with destructuring
arrays vs objects (need to use [] vs {}). For
example, Object.entries(...) returns an array,
so should be destructured with []

Also: Minor additions like total price